Specialize iterate for `IndexLinear` arrays (#58635)
This adds a specialized `iterate` method for linearly indexed arrays.
Firstly, we explicitly evaluate the iteration over the range that
corresponds to the linear indices. Secondly, we return `nothing` for an
out-of-bounds state, instead of throwing a `BoundsError`. This also lets
us annotate the actual indexing with `@inbounds`.
On master
```julia
julia> A = rand(1000,1000); v = view(A, :); v2 = view(A, 1:2:lastindex(A));

julia> using LinearAlgebra

julia> @btime norm(Iterators.map(splat(-), zip($A, $A)));
  434.982 μs (0 allocations: 0 bytes)

julia> @btime norm(Iterators.map(splat(-), zip($v, $v)));
  2.080 ms (0 allocations: 0 bytes) # v"1.13.0-DEV.709"
  510.835 μs (0 allocations: 0 bytes) # this PR
```
The performance of iterating over the contiguous `view` becomes
comparable to that of the `Array`.

Fixes https://github.com/JuliaLang/julia/issues/43295 without the need
to disable bounds-checking.

---------

Co-authored-by: N5N3 <2642243996@qq.com>