GC Always Full Flag (#58713)
This used to be an environment variable (`JULIA_GC_NO_GENERATIONAL`)
that could only be enabled if Julia was built with `WITH_GC_DEBUG_ENV`.

People benchmarking GC may want to disable generational behavior in
non-debug builds as well.

This PR introduces `gc-sweep-always-full` as a hidden
command-line-option.

It also moves the docstrings for `--hard-heap-limit` and
`--heap-target-increment` into `opts_hidden`, as it should have been
done in the original PR.