Work around LLVM JITLink stack overflow issue. (#58579)
The JITLinker recurses for every symbol in the list so limit the size of
the list

This is kind of ugly. Also 1000 might be too large, we don't want to go
too small because that wastes memory and 1000 was fine locally for the
things I tested.

Fixes https://github.com/JuliaLang/julia/issues/58229