build: Add --no-same-owner to TAR
tar changes behavior when the current uid is 0 to try to also restore
owner uids/gids (if recorded). It is possible for the uid to be 0 in
single-uid environments like user namespace sandboxes, in which case
the attempt to change the uid/gid fails. Of course ideally, the tars
would have been created non-archival (so that the uid/gid wasn't recorded
in the first place), but we get source tars from various places, so
we can't guarantee this. To make sure we don't run into trouble,
manually add the --no-same-owner flag to disable this behavior.