bpart: Properly track methods whose source was invalidated since require_world
There are three categories of methods we need to worry about during staticdata
validation:
1. New methods added to existing generic functions
2. New methods added to new generic functions
3. Existing methods that now have new CodeInstances

In each of these cases, we need to check whether any of the implicit
binding edges from the method's source was invalidated. Currently,
we handle this for 1 and 2 by explicitly scanning the method on load.
However, we were not tracking it for case 3. Fix that by using an
extra bit in did_scan_method that gets set when we see an existing
method getting invalidated, so we know that we need to drop the
corresponding CodeInstances during load.

Fixes #58346