Make current_terminfo a OncePerProcess
There seems to be no reason to always load this unconditionally - especially since
it's in the critical startup path. If we never print colored output or our IO
is not a TTY, we don't need to load this at all. While we're at it, remove the
`term_type` argument to `ttyhascolor`, which didn't work as advertised anyway,
since it still looked at the current_terminfo. If clients want to do a full
TermInfo check, they can do that explicitly.