Base: Add `count` argument to `Base.summarysize`
Frequently the number of allocations can be more important for
performance than the amount of memory allocated.

This adds a small utility to `summarysize` to allow counting the number
of unique reachable objects, which can be useful to understand, e.g.,
if you are performing many small allocations or how many of the
allocations in a function ended up reachable from the result.