bpart: Tweak `isdefinedglobal` on backdated constant
In d2cc06193ef4161e4ac161bd4b5b57a51686a89a and prior commits, we
made backdated access a conditional error (if depwarns are enabled
or in generators). However, we did not touch `isdefinedglobal`.
This resulted in the common pattern `isdefinedglobal(m, s) && getglobal(m, s)`
to sometimes error. In particular, this could be observed when attempting
to print a type from inside a generated function before that type's
definition age.

Additionally, I think the usage there, which used `invokelatest` on each
of the two queries is problematic because it is racy, since the two
`invokelatest` calls may be looking at different world ages.

This makes two tweaks:
1. Makes `isdefinedglobal` consistent with `getglobal` in that it now returns
   false if `getglobal` would throw due to the above referenced restriction.
2. Removes the implicit `invokelatest` in _isself in the show code.
   Instead, it will use the current world. I considered having it use
   the exception age when used for MethodErrors. However, because this is
   used for printing it matters more how the object can be accessed *now*
   rather than how it could have been accessed in the past.