`StepRangeLen` - conversion not contruction on indexed access
Conversion to element type `T` is implied for indexed access to `StepRangeLen` range,
from the result of arithmetic on range parameters `(ref::R, step::S, len::L, offset::L)`
and the subscript.

Previous method assumes constructor `T(r::R2)` was available for this conversion,
which made indexed access break when no such constructor was available.

This change is to use conversion rather than construction,
which is more likely to be available, and follows the guidance in
the manual [Conversion vs. Construction](https://docs.julialang.org/en/v1/manual/conversion-and-promotion/#Conversion-vs.-Construction)
considering that the conversion is implicit.