Fix use-after-free in FileWatching
We observe an abort on Windows on Revise master CI, where a free'd handle
is passed to jl_close_uv. This PR appears to fix that issue, though I have
yet to understand the full sequence of events that is causing this issue
in order to know whether any further changes are required.