WIP/RFC: Add abi tags for ucrt/cxxlib
This is a prepratory commit for reflecting windows/mingw32 libc information
in our binary platforms, as well as adding the basic detection for ucrt in
our Makefiles (so it doesn't accidentally try to use binaries with the
wrong ABI). Of course there are policy choices to be made about what we want
to ship and test. This does not make any of those policy choices, it simply
aims to provide the capability to represent these ABIs in case we decide to
add them later.

To reflect the possibility of a libc++ difference (mingw comes in both libc++
and libstdc++ flavors these days and the same is true for linux of course,
although less prominently so), this adds a new `cxxlib` abi tag that can be
either `libstdcxx` or `libcxx` (perhaps `msvc` in the future, since clang does
support that). There is also `cxxlib_version`. For compatibility the existing
`libstdcxx_version` implies both `cxxlib=libstdcxx` as well as the corresponding
versions. I'm not weeded to this representation, I just figured it was a reasonable
first attempt.

Fixes #59009