[REPL] Handle empty completion, keywords better
When the context is empty, (like "<TAB><TAB>"), return only names local to the
module (fixes #58931).

If the cursor is on something that "looks like" an identifier, like a boolean or
one of the keywords, treat it as if it was one for completion purposes.  Typing
a keyword and hitting tab no longer returns the completions for the empty
input (fixes #58309, #58832).