build: Also pass -fno-strict-aliasing for C++
As diagnosed by Andrew Pinski (https://github.com/JuliaLang/julia/issues/58466#issuecomment-3105141193),
we are respecting strict aliasing currently. We turn this of for C, but
the flag appears to be missing for C++. Looks like it's been that
way ever since that flag was first added to our build system (#484).
We should probably consider running TypeSanitizer over our code base
to see if we can make our code correct under strict aliasing as
compilers are increasingly taking advantage of it.

Fixes #58466