inference: Make test indepdent of the `Complex` method table
The test helpers in both the base tests and linear algebra add
`(::Type{<:Number})(::Furlong)` methods (for their own respective
copies of furlong). These methods intersect with `Complex(::Any)`.
Having only one of these isn't a problem, but if both get added,
we consider that too many methods and bail. In discussion this
with Jeff (who originally wrote the test), our preference is
that inference tests should be as standalone from the base
system's method tables as possible, so adjust the test (as
opposed to e.g. changing the Furlong signatures).

Fixes #59073