build: add --fix option for check-whitespace
Help agents (and humans) to solve whitespace quicker. New options
include  to try fixing issues before checking for errors that cannot be
automatically fixed (tabs in the middle of lines) and the ability to
pass specific files as arguments (particularly for testing this script
or for files that aren't yet in git).