WIP/RFC: Change lowering of GC.preserve to be lexical
# Disclaimer

This is an attempt to resolve #59124. It is incompletely implemented, but I think
the key pieces are implemented in each of the involved subsystems. The idea of
this PR is to facilitate discussion of this solution while making sure there aren't
any major unknown unknowns (which I don't think there are).

# Design

This PR adds a new struct defined as:
```
struct GCPreserveDuring
    f::Any
    # N.B: This field is opaque - the compiler is allowed to arbitrarily change it
    # as long as it has the same GC rooting behavior.
    root::Any
    GCPreserveDuring(@nospecialize(f), @nospecialize(root)) = new(f, root)
end

# This has special support in inference and codegen and is only ever actually called
# in fallback cases.
function (this::GCPreserveDuring)(args...)
    @noinline
    r = this.f(args...)
    # N.B.: This is correct, but stronger than required. If the call to `f` is deleted,
    # this may be deleted as well.
    donotdelete(this.root)
    return r
end
```

The idea is that the call method here exists for semantics, but is essentially never
actually used. Instead, inference treats it transparently (as if a call to the wrapped
function) and codegen codegens it as the original call plus a `jl_roots` operand bundle
(which our LLVM passes already supported, because it's used in the gc preserve implementation
for foreigncall).

# Key notes for relevant subsystems

## Lowering

In lowering, the `gc_preserve` syntax form is expanded by first expanding its first
argument and then rewriting every call in the expansion from `(call f ,@args)` to
`(call (new (top GCPreserveDuring) f preservee) ,@args)`. As lowering is lexical,
this will apply to every call in the lexical scope of the macro (and no other calls).

Now, of course the preservee itself is not lexical. Lowering itself does not care and
simply copies whatever value you put there. However, to preserve the existing semantics
of the `@GC.preserve` macro, the implementation of the macro has changed to introduce
and intermediate slot (thus capturing the value of the preserved slot at entry to the macro).

As a result, the case where the entry to the macro is not evaluated for any reason now
errors with an UndefVarError, rather than crashing in the optimizer.

## Inlining

Inlining for calls of `GCPreserveDuring` is adjusted to apply the same `GCPreserveDuring`
to every statement being inlined.

## :invoke_modify

The `:invoke_modify` expr head gains a new case for invokes of `GCPreserveDuring`, which
if recognized as such are treated the same as regular `:invoke`s, except that the preservation
is applied. I should note though that this causes problems for existing `:invoke_modify`',
since there is a semantic ambiguity of whether they are permitted to be decayed to `:invoke`.
Currently this is handled by just bailing in this case, but we may want a new expr head instead.

## Codegen

Codegen is the least implemented. In the current design, the `jl_cgval_t`
struct gains a new `wrapped_typ` field that if set indicates that this is
a GCPreserveDuring of the indicated type. `boxed` would reconstitute it
as such and the various `emit_` wrappers would turn it back into a non-wrapped
`jl_cgval_t` and treat the preservees as appropriate. I'm not sure this'll
be the final design.

In any case, the preservees are turned into jl_roots operand bundles.