build: Fix libssh2 source build
Addresses the same CI failure that https://github.com/JuliaLang/julia/pull/59141
was intended to address. The issue here is twofold. First, in a source
build, we failed to install openssl header files. Second, even if we
had done this, libssh needs to be explicitly told where to find openssl,
otherwise it might prefer systme files.