upgrade rapidhash v1 to v3 and fix several bugs
The rapidhash v3 algorithm supports streaming data, which makes it much
more suitable for use as a string hasher than the original. This is
actually v3 nano, since it is more similar to the original v1, but
without as much vectorization performance improvements v3.

It also fixes several bugs in the translation to hashing integers.

Initial version written by Claude, with full review and correction by
myself afterwards (it caught none of the bugs).