Try to remove `readavailable` sections from REPL precompile
The `readavailable` is problematic, because it can read partial error
messages. The `readuntil`s should be safe because they properly fence
between each prompt. If there is some other reason these readavailables
are required, let's find out and replace them with proper readuntils
if we can.

Fixes #59099 (or tries to at least)