Require `n::Integer` for `sizehint!(a, n)` (#59168)
Currently we're a bit hit-and-miss about any type-restriction on the
second argument to `sizehint!`: of 11 methods, 4 require `n::Integer`
and the remainder are unrestricted. Arguably, our dispatch is "prettier"
if methods differ only in their first argument and are consistent about
type-restrictions on `n`.

This imposes `::Integer` restrictions on all methods that specify a
non-abstract type for the first argument. Only the fallback methods for
`AbstractDict`, `AbstractSet`, and `AbstractArray` are left
unrestricted. It also makes similar changes to a `rehash!` and
`hashindex` method.

On 1.12-rc1, JET shows 40 runtime dispatches for `report_opt(merge!,
(AbstractDict, AbstractDict))`, most of which are triggered from a
branch on `sizehint!(::WeakKeyDict, ::Any)`. This seems likely to
silence most of those, but it's not possible to check due to JET
currently throwing errors on master.

Because of these unrestricted fallbacks, there is one funny thing which
applies to both this PR and master:
```julia
julia> v = [1, 2, 3];

julia> sizehint!(v, 7.0);   # no-op
```
This is mitigated by the fact that `sizehint!` is just a hint. To me it
still seems like it might be better to throw an error in such cases, but
that is not implemented by this PR.

---------

Co-authored-by: Shuhei Kadowaki <aviatesk@gmail.com>