more accurate `rad2deg` and `deg2rad` for `Float16` and `Float32` (#59097)
Both `deg2rad(::AbstractFloat)` and `rad2deg(::AbstractFloat)` just
multiply the input argument with a constant-foldable factor of the same
type as of the input argument.

The generic code for computing the factor involves a double rounding, so
the factor is not guaranteed to be correctly rounded.

Introduce special cases for `Float16` and `Float32` to make the factor
correctly rounded. This is not necessary for `Float64`, as the factor
happens to be correctly rounded in that case.

Direct improvements:

* improve accuracy of `deg2rad(::Float16)`

* improve accuracy of `rad2deg(::Float16)`

* improve accuracy of `rad2deg(::Float32)`

Example indirect improvements:

* improving accuracy of `rad2deg(::Float32)` also improves accuracy of
`asind(::Float32)`, `acosd(::Float32)`, `atand(::Float32)`