Improving build instructions for RHEL derivatives (#59213)
For a start, I've only added the package names for `rpm`-based
distributions and fixed a wrong number. If you want to, you can merge it
right away. But I can also try and see whether there are more hints and
docs that I can update in the following week(s).