compiler: consolidate staticdata.jl and invalidations.jl into Compiler proper
We actually might want a `Base.CompilerUtils` module so that the
`Compiler` package itself is smaller and less reliant on internal APIs,
but for now just move these into a submodule of Compiler.

These are supposed to call back into (and be called by) inference quite
heavily in the future, so it makes more sense (and is necessary for the
world-age) for them to actually be part of inference.