LazyLibrary: Make `add_dependency!` always ephemeral to a given process
Like many registration / hook-like functions in Julia, this function is
somewhat dangerous to use at compile-time. Depending on whether the
`LazyLibrary` you are adding dependencies to was constructed in a prior
pre-compilation process or not, your dependencies may or may not persist.

This change errs on the side of consistency, and makes it so that
`add_dependency!` is always an ephemeral operation that does not persist
beyond the lifetime of the process. This makes it safe to temporarily
register dependencies, e.g., during a sysimage build.